fix: remove malformed file causing deployment failure
Delete erroneous file to resolve syntax error and allow successful build.

+++ lib/utils.ts
@@ -15,8 +15,8 @@ export function formatCurrency(amount: number): string {
 export function formatDate(date: Date | string): string {
   const d = typeof date === "string" ? new Date(date) : date
   return new Intl.DateTimeFormat("en-US", {
-    year: "numeric",
-    month: "long",
+    year: "short",
+    month: "short",
     day: "numeric",
   }).format(d)
 }
